Se agrego el endpoint para actualizar la historia de usuario por sprint, aun falata hacer el sevicio para que sea funcional

+++ src/main/java/com/wposs/scrum_back/sprintuserstory/controller/SprintUserstoryController.java
@@ -1,12 +1,10 @@
 package com.wposs.scrum_back.sprintuserstory.controller;
 
 import com.wposs.scrum_back.Exception.exceptions.MethodArgumentNotValidException;
-import com.wposs.scrum_back.sprint.dto.SprintDtoRequest;
-import com.wposs.scrum_back.sprintemployee.dto.SprintEmployeeDtoRequest;
 import com.wposs.scrum_back.sprintuserstory.dto.SprintUserstoryDto;
 import com.wposs.scrum_back.sprintuserstory.dto.SprintUserstoryDtoRequest;
 import com.wposs.scrum_back.sprintuserstory.service.SprintUserstoryService;
-import com.wposs.scrum_back.userstory.service.UserStoryService;
+import com.wposs.scrum_back.taskteam.dto.TaskTeamDto;
 import com.wposs.scrum_back.utils.JWTUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -21,11 +19,9 @@ import javax.validation.Valid;
 import java.util.List;
 import java.util.UUID;
 
-
 @RestController
 @RequestMapping("/sprintuserstory")
 public class SprintUserstoryController {
-
     @Autowired
     private SprintUserstoryService sprintUserstoryService;
     @Autowired
@@ -78,6 +74,26 @@ public class SprintUserstoryController {
 
         }
     }
+    @PutMapping("/updatesprintuserstory/{idSprint}/{idUserStory}")
+    @Operation(summary = "Update sprintuserstory")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200",description = "Update sprintuserstory Success"),
+            @ApiResponse(responseCode = "400",description = "Update sprintuserstory Bad Request")
+    })
+    public ResponseEntity<SprintUserstoryDto> updateTaskTeam(@Valid @RequestBody SprintUserstoryDto sprintUserstoryDto,@PathVariable("idUserStory") UUID idUserStory,@PathVariable("idSprint") UUID idSprint, BindingResult result, @RequestHeader(value="Authorization") String token){
+        try{
+            if(jwtUtil.getKey(token) != null) {
+                if (result.hasErrors()){
+                    throw new MethodArgumentNotValidException(result.getFieldError().getDefaultMessage()+" usted ingreso: "+result.getFieldError().getRejectedValue(),"400",HttpStatus.BAD_REQUEST);
+                }
+                return new ResponseEntity<>(sprintUserstoryService.updateUserstoryService(idSprint,idUserStory),HttpStatus.OK);
+            }
+            return ResponseEntity.badRequest().build();
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
+        }
+    }
 
     @DeleteMapping("/deletesprintuserstory/{idSprint}/{idUserStory}")
     @Operation(summary = "Delete sprintuserstory")
@@ -86,11 +102,22 @@ public class SprintUserstoryController {
             @ApiResponse(responseCode = "404",description = "sprintuserstory Not Found")
     })
     public ResponseEntity deleteTaskTeam(@PathVariable("idUserStory") UUID idUserStory,@PathVariable("idSprint") UUID idSprint, @RequestHeader(value="Authorization") String token){
-
-                if (sprintUserstoryService.deleteSpringUserStory(idSprint,idUserStory)){
-                    return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        try{
+            if(jwtUtil.getKey(token) != null) {
+                try{
+                    if (sprintUserstoryService.deleteSpringUserStory(idSprint,idUserStory)){
+                        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+                    }
+                    return new ResponseEntity<>(HttpStatus.NOT_FOUND);
                 }
-                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                catch (Exception e){
+                        return ResponseEntity.badRequest().build();
+                    }
+                }
+                return ResponseEntity.badRequest().build();
+            }catch (Exception e){
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 
+            }
     }
 }
